Add a nodeSelector strategy
Presently there is only the node affinity strategy which checks node
label constraints, while in practice, pods may be constrained by either
of the affinity or nodeSelector terms. We therefore add a new strategy
that carries out label checking directly on the nodeSelector terms.

For anyone wishing a comprehensive eviction strategy based off of sudden
label non-existence, this should be used together with the nodeAffinity
strategy to ensure that both cases are caught.