add strategy defaults and fix nodeFit for node affinity strategy
This adds defaults for strategies. Now each strategy may have
different default parameters, if some are omitted from the configmap.

Since the NodeFit parameter was a bool instead of a pointer, it
couldn't be detected whether it was being set in the configmap, or
not. Hence it is changed to a pointer. If it is not set in the
configmap, the default is taken.

Except for node affinity, all other  strategies are now having a
default of false for an omitted NodeFit parameter, which matches
exactly how things have worked.

For the node affinity strategy, the legacy mode of operation has
always been what a "true" NodeFit would be. Thus an omitted NodeFit
parameter for that strategy is set to true, which matches how
node affinity has been working, effectively not evicting anything
if the pod doesn't fit into another node.

The change is in how things work when the NodeFit is explicitly set
to "false". The hard-coded check for pod fitting any node is removed,
and the NodeFit feature has control. NodeFit being false, node
affinity strategy will evict the pod regardless of whether there is
another node where it would fit.

Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>

pkg/descheduler/strategies/nodeutilization/highnodeutilization.go
pkg/descheduler/strategies/nodeutilization/lownodeutilization.go