Use Boolean logic for descheduler.alpha.kubernetes.io/evict annotation
In all the example and tests descheduler.alpha.kubernetes.io/evict
annotation was used with = true value to force a pod to
be considered evictable regardless of all the other checks.
Surprisingly descheduler.alpha.kubernetes.io/evict=false
was also producing the same effect.
Properly implement Boolean logic so that
descheduler.alpha.kubernetes.io/evict=false can be used
to configure a pod to be ignored by the descheduler
regardless of all the other checks.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>