Deprecate descheduler.alpha.kubernetes.io/evict in favor of descheduler.beta.kubernetes.io/evict with Boolean logic
In the past the descheduler was only checking the presence
of descheduler.alpha.kubernetes.io/evict annotation to force
a pod to be an eviction candidate.
descheduler.alpha.kubernetes.io/evict=false was producing the
same effect of descheduler.alpha.kubernetes.io/evict=true.
If the user wants to exclude a specific pod from the descheduler
actions, the suggested solution was to re-create it
setting an higher priority.
This could potentially work for best effort pods that can
be easily recreated but it's not a viable solution for other use
cases like VMs (executed in pod) that can be temporarily
excluded from the descheduler action without the need to reboot
them.
Let's deprecate descheduler.alpha.kubernetes.io/evict introducing
descheduler.beta.kubernetes.io/evict with proper Boolean logic.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>